Disable modeline parsing for security
Mode lines are lines that can be added to a text file to
make Vim execute arbitrary commands when started. A
maliciously crafted file could present a security risk.

Patches were released for Vim (patch 8.1.1365) and
Neovim (patch released in v0.3.6). Since modelines are
unlikely to be used disabling them is an additional
measure.

https://github.com/numirias/security/blob/master/doc/2019-06-04_ace-vim-neovim.md
https://www.techrepublic.com/blog/it-security/turn-off-modeline-support-in-vim/